Add note for array with empty struct or union
Fix #358

The key point of this issue is C and C++ having different type size for
array of empty struct or union, size of empty struct or union is 0 for
C, and 1 for C++, and the size of array is type size multiplied by the
length of the array.

Also checked AArch64 code gen, they also generated different code between C
and C++, and ABI isn't explicitly describe this but has describe the size
of array is type * length.